+++ panel.c
@@ -5,8 +5,8 @@
 #include "a_samp.h"
 #include "game_sa.h"
 #include "panel.h"
+#include "math.h"
 #include "nav.h"
-#include <math.h>
 #include <string.h>
 
 /*
@@ -356,7 +356,7 @@ void panel_timed_update(AMX *amx)
 		x = *((float*) buf32);
 		y = *((float*) buf64);
 		z = *((float*) buf144);
-		x = VEL_TO_KTS_VAL * sqrtf(x * x + y * y + z * z);
+		x = VEL_TO_KTS_VAL * sqrt(x * x + y * y + z * z);
 		panel_update_speed(amx, playerid, (int) x);
 
 		z *= VEL_TO_KFPM_VAL * 14.5f;
